% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fntcs.R
\name{calculate_TOP_ART_score}
\alias{calculate_TOP_ART_score}
\title{calcuation of TOP-ART score}
\usage{
calculate_TOP_ART_score(
  SAMPLE_ID = NULL,
  SEQ_METHOD,
  HRD,
  LST,
  PURITY,
  PLOIDY,
  SEX,
  FILE_BAM,
  FILE_RNA_BAM,
  FILE_VCF,
  GR_CNV,
  GR_SOM_SMALL_VARS = NULL,
  GR_GERM_SMALL_VARS = NULL,
  GR_GENCODE_EXON,
  GR_FULL_VARS = NULL,
  TBL_YAPSA = NULL,
  FILE_REFGEN = NULL,
  TOTAL_SNVS_YAPSA = NULL,
  GR_SOM_SNV_YAPSA = NULL,
  PRINT_OUTPUT = F,
  STORE_OUTPUT = F,
  OUTPUT_DIR = NULL,
  MAN_SOM_GENES = NULL,
  MAN_GERM_GENES = NULL,
  DEBUG = F,
  FILTER_SOM_SMALL_VARS = T,
  showReadDetail = F
)
}
\description{
calcuation of TOP-ART score
}
