% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fntcs.R
\name{estimate_signature_score}
\alias{estimate_signature_score}
\title{calculate signature score}
\usage{
estimate_signature_score(
  TBL_YAPSA = NULL,
  SEQ_METHOD,
  run_yapsa = F,
  FILE_REFGEN = NULL,
  TBL_SOM_SNV_YAPSA = NULL,
  TOTAL_SNVS_YAPSA = NULL
)
}
\description{
calculate signature score
}
