% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fntcs.R
\name{run_YAPSA_signature_analysis}
\alias{run_YAPSA_signature_analysis}
\title{run YAPSA analysis of mutational signatures}
\usage{
run_YAPSA_signature_analysis(FILE_REFGEN, SEQ_METHOD, TBL_SOM_SNV_YAPSA)
}
\arguments{
\item{FILE_REFGEN}{indexed fasta file of reference genome .fa}

\item{SEQ_METHOD}{character either "WES" or "WGS".}

\item{TBL_SOM_SNV_YAPSA}{tibble/dataframe containing relevant SNVs for YAPSA
analysis.}
}
\value{
A tibble containing the results for each called signature.
}
\description{
run YAPSA analysis of mutational signatures
}
